updating threads and filtering

+++ 03_sort.sh
@@ -2,6 +2,7 @@
 
 module add UHTS/Analysis/samtools/1.8;
 SRR_exp=$1
+nThreads=$2
 FILES=$(find $working_path/$SRR_exp/bam/ -type f -name "*.bam")
 echo $FILES
 for f in $FILES
@@ -9,10 +10,13 @@ do
   echo "Sorting $f..."
   target_name=${f##*/}
   target_name=${target_name%.bam}
+  target_name=${target_name%_sorted}
 #  echo $target_name
-  samtools sort $f > $working_path/$SRR_exp/bam/${target_name}_sorted.bam
-  echo "Indexing $f..."
-  samtools index $working_path/$SRR_exp/bam/${target_name}_sorted.bam
-  rm $f
+  if [ ! -f "$working_path/$SRR_exp/bam/${target_name}_sorted.bam" ]; then
+    samtools sort -@ $nThreads $f > $working_path/$SRR_exp/bam/${target_name}_sorted.bam
+    echo "Indexing $f..."
+    samtools index -@ $nThreads $working_path/$SRR_exp/bam/${target_name}_sorted.bam
+    rm $f
+  fi
 done
 module rm UHTS/Analysis/samtools/1.8;
